data_processing/requirements.txt
pandas==1.5.3
numpy==1.24.3
requests==2.31.0
tqdm==4.65.0
pathlib==1.0.1
python-dotenv==1.0.0
openai==1.3.0
argparse==1.4.0
httpx==0.24.1
pillow==9.5.0
rich==13.4.2